refactor: sharpen workspace package boundaries
Move repository membership and status rendering into focused files, place shared helpers with their owning concerns, and remove the obsolete prepared-claim spike test.
